[NTOSKRNL] - Make file offsets 64 bit - Fix lots of MSVC/64 bit warnings
svn path=/trunk/; revision=56264

+++ reactos/ntoskrnl/cache/section/reqtools.c
@@ -61,7 +61,7 @@ MiGetOnePage(PMMSUPPORT AddressSpace,
              PMEMORY_AREA MemoryArea,
              PMM_REQUIRED_RESOURCES Required)
 {
-    int i;
+    ULONG i;
     NTSTATUS Status = STATUS_SUCCESS;
 
     for (i = 0; i < Required->Amount; i++)
